refactor(apps): trim native skills review surface
